requirements.txt
aiobotocore==2.11.2
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.32.49
boto3==1.34.50
botocore==1.34.51
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.4
docutils==0.16
frozenlist==1.4.1
fsspec==2024.2.0
idna==3.6
jmespath==1.0.1
multidict==6.0.5
numpy==1.26.4
pandas==2.2.1
pyasn1==0.5.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
s3fs==2024.2.0
s3transfer==0.10.0
six==1.16.0
tzdata==2024.1
urllib3==2.0.7
wrapt==1.16.0
yarl==1.9.4

